endpoint for attachments doctor's medical exams return all exams from the same department

+++ src/main/java/com/dm/MedicalDocumentation/prescription/PrescriptionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PrescriptionRepository extends JpaRepository<Prescription, Integer> {
+public interface PrescriptionRepository extends JpaRepository<Prescription, Long> {
     List<Prescription> findByPatientUserUserLogin(String userLogin);
 }
